requirements.txt
flask
requests
pandas
joblib
pillow
scikit-learn
